refactor: use absolute paths in @/src/**
