+++ src/pages/index.test.tsx
@@ -3,10 +3,9 @@ import "isomorphic-unfetch"
 import { rest } from "msw"
 import { setupServer } from "msw/node"
 
+import App from "@/src/pages/index"
 import { render, screen } from "@testing-library/react"
 import userEvent from "@testing-library/user-event"
-
-import App from "./index"
 
 const server = setupServer(
   rest.get("https://api.openweathermap.org/*", (req, res, ctx) => {
